refactor: decompoe componente Protocolos em View e subcomponentes
- Substitui o componente monolítico Protocolos.vue por ProtocolosView.vue
- Extrai a lógica de listagem para components/protocolos/ProtocolosLista.vue
- Extrai o formulário de cadastro para components/protocolos/ProtocolosForm.vue
- Extrai o modal de detalhes para components/protocolos/ProtocolosDetalhes.vue
- Atualiza as rotas para utilizar a nova View
- Melhora a organização do código separando responsabilidades de UI e controle de estado

Fixes #10
